web/requirements.txt
datasets==2.7.1
click==8.1.3
Flask==2.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.0.0
MarkupSafe==2.1.1
Werkzeug==2.0.0
joblib==1.2.0
scikit-learn==1.1.3
psycopg2-binary==2.9.5
locust==2.13.0
python-dotenv==0.21.0